docker/alarmes/serializer.py
from rest_framework import serializers
from alarmes.models import Evento, Incidente


class EventoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Evento
        fields = '__all__'


class IncidenteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Incidente
        fields = '__all__'
