docker/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
joblib==1.0.0
Markdown==3.3.3
nltk==3.5
numpy==1.19.4
pandas==1.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.55.0
urllib3==1.26.2
